+++ bindings/go/evmc/host.c
@@ -83,8 +83,8 @@ static inline void go_exported_functions_type_checks()
     size = copyCode(context, address, size, data, size);
 
     evmc_selfdestruct_fn selfdestruct_fn = NULL;
-    selfdestruct_fn(context, address, address);
-    selfdestruct(context, address, address);
+    bool_flag = selfdestruct_fn(context, address, address);
+    bool_flag = selfdestruct(context, address, address);
 
     evmc_call_fn call_fn = NULL;
     result = call_fn(context, message);